<No Comment Entered>

+++ org/scec/sha/earthquake/rupForecastImpl/PEER_TestCases/PEER_NonPlanarFaultForecast.java
@@ -274,7 +274,7 @@ public class PEER_NonPlanarFaultForecast extends EqkRupForecast
        double totMoRate = 3e10*faultArea*slipRate;
        grMagFreqDist.setAllButTotCumRate(GR_MAG_LOWER, magUpper, totMoRate,GR_BVALUE);
 
-System.out.println("rate³5="+grMagFreqDist.getCumRate(5.05));
+System.out.println("rate³5="+(float)grMagFreqDist.getCumRate(5.05));
 
        // Now make the source
        source = new  PEER_FaultSource(grMagFreqDist, RAKE ,
@@ -282,6 +282,7 @@ System.out.println("rate³5="+grMagFreqDist.getCumRate(5.05));
                                         (EvenlyGriddedSurface)surface,
                                         timeSpan.getDuration(),
                                         ((Double)lengthSigmaParam.getValue()).doubleValue() );
+
      }
      parameterChangeFlag = false;
    }
